+++ Game/Weapon/IWeapon.cs
@@ -8,7 +8,7 @@ public interface IWeapon
     int Cost { get; }
     int Damage { get; }
     //в 4 лабе distance
-    double Distance { get; }
+    int Distance { get; }
     
    
 }
