+++ Game/Weapon/Staff.cs
@@ -1,10 +1,10 @@
 namespace Game.Weapon;
 
-public class Staff(string name, int cost, int baseDamage, double specialChance, double distance)
+public class Staff(string name, int cost, int baseDamage, double specialChance, int distance)
     : IWeapon
 {
     //в 4 лабе distance
-    public double Distance { get; } = distance;
+    public int Distance { get; } = distance;
     private readonly int _baseDamage = baseDamage;
     public string Name { get; } = name;
     public int Cost { get; } = cost;
