+++ Game/Weapon/Sword.cs
@@ -1,13 +1,13 @@
 namespace Game.Weapon;
 
-public class Sword(string name, int cost, int baseDamage, double distance) : IWeapon
+public class Sword(string name, int cost, int baseDamage, int distance) : IWeapon
 {
     private readonly int _baseDamage = baseDamage;
     public string Name { get; } = name;
     public int Cost { get; } = cost;
     public int Damage { get; } = baseDamage;
     //в 4 лабе distance
-    public double Distance { get; } = distance;
+    public int Distance { get; } = distance;
 
     public double GenerateDamage()
     {
